chore: move tool schema to mcp as it is used by all servers (#887)
