+++ src/loopTools/DEPS.list
@@ -2,5 +2,4 @@
 ../
 ../loop/
 ../mcp/
-../tools/
 ../utils/
